Split inquiry status checker into own component
